user create func moved to utils
